Fix Battery light reset action bar button
This fixes an issue in where the color pickers for Battery low,
Charging and Charged (90%) aka the top 3 were not reset back to
a working state after toggling the 'Only light up when fully charged'
switch and then the reset button in the action bar.

- Remove pulse preference from being reset. This should not be reset
at all. If the user wants to change the state, they can on their own.

Change-Id: Id0f04a852894f325c9baaffa57ecc7d64301527c